requirements.txt
trax==1.3.7
pytest
requests
absl-py
pillow
matplotlib
scikit-image
